+++ lib/pages/Homepage/ForgotPasswordPage.dart
@@ -1,0 +1,38 @@
+import 'package:flutter/material.dart';
+import 'package:supabase_flutter/supabase_flutter.dart';
+
+class ForgotPasswordPage extends StatelessWidget {
+  const ForgotPasswordPage({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    final TextEditingController _emailController = TextEditingController();
+    final _supabase = Supabase.instance.client;
+
+    Future<void> _resetPassword() async {
+      if (_emailController.text.isEmpty) {
+        ScaffoldMessenger.of(context).showSnackBar(const SnackBar(content: Text('Enter your email to reset password.')));
+        return;
+      }
+
+      await _supabase.auth.resetPasswordForEmail(_emailController.text);
+      ScaffoldMessenger.of(context).showSnackBar(
+        const SnackBar(content: Text('Check your email for password reset link.')),
+      );
+    }
+
+    return Scaffold(
+      appBar: AppBar(title: const Text('Forgot Password')),
+      body: Padding(
+        padding: const EdgeInsets.all(16.0),
+        child: Column(
+          children: [
+            TextField(controller: _emailController, decoration: const InputDecoration(labelText: 'Email')),
+            const SizedBox(height: 20),
+            ElevatedButton(onPressed: _resetPassword, child: const Text('Reset Password')),
+          ],
+        ),
+      ),
+    );
+  }
+}
