+++ lib/pages/Trainerpage/button2.dart
@@ -1,0 +1,42 @@
+import 'package:flutter/material.dart';
+import 'package:supabase_flutter/supabase_flutter.dart';
+
+class button2 extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text('Gym Shop'),
+        backgroundColor: Colors.indigo[600],
+      ),
+      body: Center(
+        child: Padding(
+          padding: const EdgeInsets.all(16.0),
+          child: Column(
+            mainAxisAlignment: MainAxisAlignment.center,
+            children: [
+              Text(
+                'Gym Shop Page',
+                style: TextStyle(fontSize: 24, fontWeight: FontWeight.bold),
+              ),
+              SizedBox(height: 20),
+              Text(
+                'Here you can shop for gym equipment, merchandise, etc.',
+                textAlign: TextAlign.center,
+                style: TextStyle(fontSize: 16),
+              ),
+              SizedBox(height: 30),
+              ElevatedButton(
+                onPressed: () {
+                  // Handle shop action here
+                  print('Shopping Started');
+                },
+                child: Text('Go Shopping'),
+              ),
+            ],
+          ),
+        ),
+      ),
+    );
+  }
+}
